Ubicaciones and Tablas delete and update method

+++ code/poc/poc-persistence/src/main/java/mx/com/prosa/poc/persistence/TblTablasRepository.java
@@ -9,6 +9,6 @@ import mx.com.prosa.poc.model.TblTablas;
 /**
  * The Interface TblTablasRepository.
  */
-public interface TblTablasRepository extends JpaRepository<TblTablas, BigDecimal> {
+public interface TblTablasRepository extends JpaRepository<TblTablas, Long> {
 
 }
